[Red] Add LoginFormContainer
