+++ src/LoginPage.jsx
@@ -1,27 +1,12 @@
 import React from 'react';
-import { useSelector } from 'react-redux';
 
-import LoginForm from './LoginForm';
-
-import { get } from './utils';
+import LoginFormContainer from './LoginFormContainer';
 
 export default function LoginPage() {
-  const loginFields = useSelector(get('loginFields'));
-
-  function handleChange() {
-    //
-  }
-
-  function handleSubmit() {
-    //
-  }
   return (
     <>
-      <LoginForm
-        onChange={handleChange}
-        onSubmit={handleSubmit}
-        fields={loginFields}
-      />
+      <h2>Log In</h2>
+      <LoginFormContainer />
     </>
   );
 }
